# LogArtApp/backend/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: LogArt API
  version: 1.0.0
  description: Documentación API REST para la aplicación LogArt
  contact:
    name: David Moreno Martín
    email: david.moreno.m.w@gmail.com
    url: https://github.com/codeurjc-students/2024-logart
servers:
  - url: https://localhost:443/api/v1
    description: Servidor de desarrollo
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ObjectCreated:
      type: object
      properties:
        name:
          type: string
          example: Objeto de Ejemplo
        description:
          type: string
          example: Descripción del objeto de ejemplo.
        imageUrl:
          type: string
          example: https://example.com/images/objects/objeto-ejemplo.jpg
        discipline:
          type: string
          example: 60d0fe4f5311236168a109cb
        createdBy:
          type: string
          example: 60d0fe4f5311236168a109ca
        _id:
          type: string
          example: 60d0fe4f5311236168a109cc
        createdAt:
          type: string
          format: date-time
          example: '2022-01-01T12:34:56Z'
        __v:
          type: integer
          example: 0
    ObjectUpdated:
      type: object
      properties:
        _id:
          type: string
          example: 60d0fe4f5311236168a109cc
        name:
          type: string
          example: Objeto Actualizado
        description:
          type: string
          example: Descripción actualizada del objeto.
        imageUrl:
          type: string
          example: https://example.com/images/objects/objeto-actualizado.jpg
        discipline:
          type: string
          example: 60d0fe4f5311236168a109cb
        createdBy:
          type: string
          example: 60d0fe4f5311236168a109ca
        createdAt:
          type: string
          format: date-time
          example: '2022-01-01T12:34:56Z'
        __v:
          type: integer
          example: 0
    ObjectGetGalleryByDiscipline:
      type: object
      properties:
        _id:
          type: string
          example: 60d0fe4f5311236168a109cc
        name:
          type: string
          example: Objeto de Ejemplo
        description:
          type: string
          example: Descripción del objeto de ejemplo.
        imageUrl:
          type: string
          example: https://example.com/images/objects/objeto-ejemplo.jpg
        discipline:
          type: string
          example: 60d0fe4f5311236168a109cb
        createdBy:
          type: object
          properties:
            id:
              type: string
              example: 60d0fe4f5311236168a109ca
            username:
              type: string
              example: johndoe
            firstName:
              type: string
              example: John
            lastName:
              type: string
              example: Doe
        createdAt:
          type: string
          format: date-time
          example: '2022-01-01T12:34:56Z'
        __v:
          type: integer
          example: 0
...
    CreateObjectResponse:
      type: object
      properties:
        object:
          $ref: '#/components/schemas/ObjectCreated'
        message:
          type: string
          example: Object created successfully
...
    UpdateObjectResponse:
      type: object
      properties:
        object:
          $ref: '#/components/schemas/ObjectUpdated'
        message:
          type: string
          example: Object updated successfully
    DeleteObjectResponse:
      type: object
      properties:
        message:
          type: string
          example: Object deleted successfully
    GetGalleryByDisciplineResponse:
      type: object
      properties:
        discipline:
          type: object
          properties:
            id:
              type: string
              example: 60d0fe4f5311236168a109cb
            name:
              type: string
              example: Canciones
        totalObjects:
          type: integer
          example: 1
        objects:
          type: array
          items:
            $ref: '#/components/schemas/ObjectGetGalleryByDiscipline'
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 1
    GetObjectByIdResponse:
      type: object
      properties:
        object:
          $ref: '#/components/schemas/ObjectGetGalleryByDiscipline'
...
security:
  - BearerAuth: []
paths:
...
  /objects:
    post:
      summary: Crear un nuevo objeto
      tags:
        - Objetos
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Objeto de Ejemplo
                description:
                  type: string
                  example: Descripción del objeto de ejemplo.
                disciplineName:
                  type: string
                  example: Canciones
                imageUrl:
                  type: string
                  format: binary
                  description: Imagen del objeto
      responses:
        '201':
          description: Objeto creado exitosamente
          headers:
            Location:
              description: URL del nuevo objeto
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateObjectResponse'
        '400':
          description: Error, solicitud incorrecta o datos faltantes
        '404':
          description: Error, disciplina no encontrada
        '500':
          description: Error interno del servidor
  /objects/{objectId}:
    put:
      summary: Actualizar un objeto existente
      tags:
        - Objetos
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: objectId
          required: true
          schema:
            type: string
          description: ID del objeto a actualizar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Objeto Actualizado
                description:
                  type: string
                  example: Descripción actualizada del objeto.
                disciplineName:
                  type: string
                  example: Canciones
                imageUrl:
                  type: string
                  format: binary
                  description: Nueva imagen del objeto
      responses:
        '200':
          description: Objeto actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateObjectResponse'
        '400':
          description: >-
            Error, solicitud incorrecta, datos faltantes o formato de ID
            inválido
        '403':
          description: Error, no autorizado para actualizar este objeto
        '404':
          description: Error, objeto o disciplina no encontrada
        '500':
          description: Error interno del servidor
    delete:
      summary: Eliminar un objeto por su ID
      tags:
        - Objetos
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: objectId
          required: true
          schema:
            type: string
          description: ID del objeto a eliminar
      responses:
        '200':
          description: Objeto eliminado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteObjectResponse'
        '400':
          description: Error, formato de ID de objeto inválido
        '403':
          description: Error, no autorizado para eliminar este objeto
        '404':
          description: Error, objeto no encontrado
        '500':
          description: Error interno del servidor
  /objects/{disciplineName}:
    get:
      summary: Obtener la galería de objetos por disciplina
      tags:
        - Objetos
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: disciplineName
          required: true
          schema:
            type: string
          description: Nombre de la disciplina
        - in: query
          name: userId
          schema:
            type: string
          description: Filtrar objetos por ID de usuario
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Número de página para paginación
        - in: query
          name: limit
          schema:
            type: integer
            default: 3
          description: Número de objetos por página
        - in: query
          name: objectName
          schema:
            type: string
          description: Filtrar objetos por nombre
      responses:
        '200':
          description: Galería de objetos obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetGalleryByDisciplineResponse'
        '404':
          description: Error, disciplina no encontrada
        '500':
          description: Error interno del servidor
  /objects/details/{objectId}:
    get:
      summary: Obtener detalles de un objeto por su ID
      tags:
        - Objetos
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: objectId
          required: true
          schema:
            type: string
          description: ID del objeto a buscar
      responses:
        '200':
          description: Detalles del objeto obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetObjectByIdResponse'
        '400':
          description: Error, formato de ID de objeto inválido
        '403':
          description: Error, no autorizado para ver este objeto
        '404':
          description: Error, objeto no encontrado
        '500':
          description: Error interno del servidor
...